MetersAA/frontend/requirements.txt
# requirements.txt file
Django==4.0.5
wfastcgi
django-import-export
